backend/src/services/pythonAgents/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.0.20
langchain==0.1.0
langchain-google-genai==0.0.6
openai==1.3.0
google-api-python-client==2.105.0
google-generativeai==0.3.1
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.0
PyMuPDF==1.23.8
